chore: move CODE_GEN directories to new file
